Implement Feedback on fixing try/catch and code abstraction
