+++ src/controllers/userController.js
@@ -1,6 +1,9 @@
-import UserService from '../services/user.service';
 import ResponseGenerator from '../utils/response.util';
+import models from '../models';
+import GeneralUtils from '../utils/index';
+import Auth from '../middlewares/auth';
 
+const { Users } = models;
 const response = new ResponseGenerator();
 /**
  * @description Handles Users
@@ -15,20 +18,43 @@ class UserController {
    * @return {json} Returns json object
    * @memberof UserController
    */
-  static async login(req, res) {
+  static async login(req, res, next) {
     try {
-      const user = await UserService.login(req);
+      const { email, password } = req.body;
+      const user = await Users.findOne({
+        where: {
+          email
+        }
+      });
       if (user) {
-        return response.sendSuccess(
-          res,
-          200,
-          user,
-          'User successfully logged in'
-        );
+        const bycrptResponse = GeneralUtils.validate(password, user.password);
+        if (bycrptResponse) {
+          const {
+            id, isAdmin, firstname, lastname
+          } = user;
+          const token = await Auth.signJwt({
+            id,
+            isAdmin
+          });
+          const data = {
+            token,
+            id,
+            isAdmin,
+            firstname,
+            lastname,
+            email
+          };
+          return response.sendSuccess(
+            res,
+            200,
+            data,
+            'User successfully logged in'
+          );
+        }
       }
-      return response.sendError(res, 500, 'Something went wrong');
+      return response.sendError(res, 400, 'Invalid email or password');
     } catch (err) {
-      return response.sendError(res, 400, err.message);
+      next(err);
     }
   }
 }
